+++ config/routes.rb
@@ -58,9 +58,7 @@ Dorsale::Engine.routes.draw do
     resources :activity_types, except: [:destroy, :show]
     resources :origins, except: [:destroy, :show]
 
-    namespace :people do
-      get :activity
-    end
+    resources :events, only: [:index]
 
     resources :people do
       resources :links