+++ db/migrate/20170915070538_dorsale_customer_vault_events_add_text.rb
@@ -2,7 +2,7 @@ class DorsaleCustomerVaultEventsAddText < ActiveRecord::Migration[5.0]
   def change
     add_column :dorsale_customer_vault_events, :text, :text
     model = Dorsale::CustomerVault::Event
-    model.where(action: "create").update_all(text: model.t("action.create"))
-    model.where(action: "update").update_all(text: model.t("action.update"))
+    model.where(action: "create").update_all(text: model.t("text.create"))
+    model.where(action: "update").update_all(text: model.t("text.update"))
   end
 end